Rename python/sglang/multimodal_gen/runtime/pipelines_core/stages/adapter.py
