+++ python/sglang/multimodal_gen/runtime/pipelines_core/stages/text_connector.py
